Binary support in BSON (dev bundle wise) and in Mongo
Unfortunately, the Node Mongo driver wants to use its own proprietary
Binary type for binary data.  We want to use Uint8Array, instead, so
we're going to be using a patched version of the BSON library that
allows this.

It allows us to install a handler for USER_DEFINED binary types, which
in our case we set up to recognize, accept, and provide Uint8Array